change net module abstraction,
_net module:

udp_conn type own socket, port,
for listener inc packet source addr
for dial specifies destination

routines implemented:
        to listen and read
to be implemented:
        to dial and send

_server module:

sessions type has head of linked list of heads and lock
to be considered shared data.

stubs to operate sessions, init destroy

_main module:

make use of new _net abstractions, thread pool,
sessions bootstraping.

+++ thread/server/_utils.c
@@ -19,7 +19,7 @@ int8_t _parse_port(char *portstr, uintmax_t *portdest) {
         return 0;
 }
 
-void _check_err(int8_t errcode, char *msg, bool fatal) {
+void _check_err(int64_t errcode, char *msg, bool fatal) {
         if (errcode < 0)
                 fprintf(stderr, "%s\n", msg);
         if (errcode < 0 && fatal)
